+++ feature/groupie/src/main/java/com/example/groupie/GroupieFragment.kt
@@ -1,4 +1,4 @@
-package com.example.androidtechsample.ui.groupie.basic
+package com.example.groupie
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,22 +7,23 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.example.androidtechsample.databinding.FragmentGroupieBasicBinding
+import com.example.androidtechsample.util.navigator
+import com.example.groupie.databinding.FragmentGroupieBinding
 import com.xwray.groupie.GroupieAdapter
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class GroupieBasicFragment : Fragment() {
+class GroupieFragment : Fragment(), GroupieItem.Listener {
 
-  private lateinit var binding: FragmentGroupieBasicBinding
-  private val viewModel: GroupieBasicViewModel by viewModels()
+  private lateinit var binding: FragmentGroupieBinding
+  private val viewModel: GroupieViewModel by viewModels()
 
   override fun onCreateView(
     inflater: LayoutInflater,
     container: ViewGroup?,
     savedInstanceState: Bundle?
   ): View {
-    binding = FragmentGroupieBasicBinding.inflate(inflater)
+    binding = FragmentGroupieBinding.inflate(inflater)
     return binding.root
   }
 
@@ -35,12 +36,18 @@ class GroupieBasicFragment : Fragment() {
       recyclerView.layoutManager = LinearLayoutManager(context)
     }
     viewModel.itemList.observe(viewLifecycleOwner) {
-      adapter.update(it)
+      for (item in it) {
+        adapter.add(GroupieItem(item, this))
+      }
     }
   }
 
   override fun onStart() {
     super.onStart()
     viewModel.fetchData()
+  }
+
+  override fun onItemClick(id: Int) {
+    navigator(id)
   }
 }
