+++ feature/groupie/src/main/java/com/example/groupie/basic/GroupieBasicFragment.kt
@@ -1,4 +1,4 @@
-package com.example.androidtechsample.ui.groupie
+package com.example.groupie.basic
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,23 +7,22 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.example.androidtechsample.databinding.FragmentGroupieBinding
-import com.example.androidtechsample.util.navigator
+import com.example.androidtechsample.databinding.FragmentGroupieBasicBinding
 import com.xwray.groupie.GroupieAdapter
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class GroupieFragment : Fragment(), GroupieItem.Listener {
+class GroupieBasicFragment : Fragment() {
 
-  private lateinit var binding: FragmentGroupieBinding
-  private val viewModel: GroupieViewModel by viewModels()
+  private lateinit var binding: FragmentGroupieBasicBinding
+  private val viewModel: GroupieBasicViewModel by viewModels()
 
   override fun onCreateView(
     inflater: LayoutInflater,
     container: ViewGroup?,
     savedInstanceState: Bundle?
   ): View {
-    binding = FragmentGroupieBinding.inflate(inflater)
+    binding = FragmentGroupieBasicBinding.inflate(inflater)
     return binding.root
   }
 
@@ -36,18 +35,12 @@ class GroupieFragment : Fragment(), GroupieItem.Listener {
       recyclerView.layoutManager = LinearLayoutManager(context)
     }
     viewModel.itemList.observe(viewLifecycleOwner) {
-      for (item in it) {
-        adapter.add(GroupieItem(item, this))
-      }
+      adapter.update(it)
     }
   }
 
   override fun onStart() {
     super.onStart()
     viewModel.fetchData()
-  }
-
-  override fun onItemClick(id: Int) {
-    navigator(id)
   }
 }
