+++ feature/groupie/src/main/java/com/example/groupie/basic/GroupieBasicViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.androidtechsample.ui.groupie.basic
+package com.example.groupie.basic
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
