+++ feature/groupie/src/main/java/com/example/groupie/carousel/GroupieCarouselViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.androidtechsample.ui.groupie.swipe
+package com.example.groupie.carousel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
@@ -6,16 +6,22 @@ import androidx.lifecycle.ViewModel
 import com.xwray.groupie.Group
 import com.xwray.groupie.GroupieAdapter
 
-class GroupieSwipeViewModel : ViewModel() {
+class GroupieCarouselViewModel : ViewModel() {
 
   private val _itemList = MutableLiveData<List<Group>>()
   val itemList: LiveData<List<Group>> = _itemList
 
   fun fetchData(colors: IntArray, adapter: GroupieAdapter) {
     val list = mutableListOf<Group>()
-    for (i in 0..10) {
-      list += GroupieSwipeItem(colors[i])
+    for (i in 0..9) {
+      list += GroupieCarouselItem(colors[i], adapter)
     }
     _itemList.value = list
   }
+
+//  fun delete(postion: Int) {
+//    list.removeAt(postion)
+//    _itemList.value = list
+//
+//  }
 }
