+++ feature/groupie/src/main/java/com/example/groupie/swipe/GroupieSwipeViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.androidtechsample.ui.groupie.carousel
+package com.example.groupie.swipe
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
@@ -6,22 +6,16 @@ import androidx.lifecycle.ViewModel
 import com.xwray.groupie.Group
 import com.xwray.groupie.GroupieAdapter
 
-class GroupieCarouselViewModel : ViewModel() {
+class GroupieSwipeViewModel : ViewModel() {
 
   private val _itemList = MutableLiveData<List<Group>>()
   val itemList: LiveData<List<Group>> = _itemList
 
   fun fetchData(colors: IntArray, adapter: GroupieAdapter) {
     val list = mutableListOf<Group>()
-    for (i in 0..9) {
-      list += GroupieCarouselItem(colors[i], adapter)
+    for (i in 0..10) {
+      list += GroupieSwipeItem(colors[i])
     }
     _itemList.value = list
   }
-
-//  fun delete(postion: Int) {
-//    list.removeAt(postion)
-//    _itemList.value = list
-//
-//  }
 }
