+++ feature/groupie/src/main/res/layout/fragment_groupie_swipe.xml
