+++ feature/groupie/src/main/res/layout/model_groupie_basic_item.xml
