+++ feature/groupie/src/main/res/layout/model_groupie_carousel_item.xml
