+++ feature/groupie/src/main/res/layout/model_groupie_item.xml
