+++ feature/groupie/src/main/res/layout/model_groupie_swipe_item.xml
